Files
This is the utorrent file. if you want to download the iso image you have to go on the internet and download utorrent and then download the torrent file. After that you open the file with utorrent then you extract the file to get the windows 11 iso image. After you got the iso file, the name of the file supposes to be "Win11_bootstrap.iso".